[FEAT] 포트폴리오 api 수정
