+++ src/main/java/com/newstoss/member/application/in/command/WithdrawService.java
@@ -5,6 +5,8 @@ import com.newstoss.global.errorcode.UserErrorCode;
 import com.newstoss.member.domain.Member;
 import com.newstoss.member.application.out.MemberCommandPort;
 import com.newstoss.member.application.out.MemberQueryPort;
+import com.newstoss.portfolio.application.port.out.DeleteMemberPnlPort;
+import com.newstoss.portfolio.application.port.out.DeleteMemberPortfolioPort;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -16,6 +18,8 @@ import java.util.UUID;
 public class WithdrawService {
     private final MemberCommandPort memberCommandPort;
     private final MemberQueryPort memberQueryPort;
+    private final DeleteMemberPnlPort deleteMemberPnlPort;
+    private final DeleteMemberPortfolioPort deleteMemberPortfolioPort;
 
     public boolean exec(UUID memberId){
         Optional<Member> member = memberQueryPort.findById(memberId);
@@ -23,6 +27,8 @@ public class WithdrawService {
             throw new CustomException(UserErrorCode.USER_NOT_FOUND);
         }
         memberCommandPort.deleteById(memberId);
+        deleteMemberPnlPort.deleteMemberPnl(memberId); //멤버와 관련된 pnl 삭제
+        deleteMemberPortfolioPort.deletePortfolio(memberId); //멤버와 관련된 포트폴리오 삭제
         return true;
     }
 }
